pyproject.toml
[project]
name = "master-agentic-ai-52w"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "anthropic>=0.84.0",
    "chromadb>=1.5.5",
    "datasets>=4.8.4",
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-pinecone==0.2.13rc1",
    "matplotlib>=3.10.8",
    "numpy>=1.26.4,<2.0.0",
    "openai>=2.28.0",
    "pandas>=3.0.1",
    "pinecone>=7.0.0,<8.0.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "tiktoken>=0.12.0",
    "transformers>=5.4.0",
    "langchain-openai>=0.3.0",
]
